Added new pages
Added login, register, and logout pages. Tried to make the forms
intelligent so that if a user has an error on submission not all of the
form data is cleared forcing them to refill it. Also built a header and
footer that will go on all pages. Need to update styling.