Updated how people follow other people
Added a row to the index page on the top posts where a user can click
follow or unfollow when they see a post they like.  Changed the
existing following page to be a place where users can remove the other
users they are following. I now need to create a page that users can go
to and just see the posts of those they follow. Also need to add
functionality for uprooting and downvot